+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 - add `INTERVAL` as a supported `return_type` for `DuckDB::ScalarFunction`. The scalar function block can now return a `DuckDB::Interval` object and it will be written back to the result vector correctly.
 - add `UUID` as a supported `return_type` for `DuckDB::ScalarFunction`. The scalar function block can now return a UUID string and it will be written back to the result vector correctly.
 - add `DuckDB::Connection#extract` to extract multiple SQL statements as prepared statements without executing them.
+- add `DuckDB::Connection#register_logical_type` to register a custom logical type with a connection.
 
 # 1.5.1.1 - 2026-04-04